Refactoring lib/posts
Separating what is specific to markdown and adding it to
lib/markdown.js so that if we have another kind of markdown
content we can use it there

Extra:
- Fix appears in link and semantics
